Create book entity and READER_OF relationship In this change, we have create the book entity and the READER_OF relationship. The changes includes: 1. Create book entity. 2. Create relationship READER_OF in Person entity to books 3. Create BookRepositoryTest to demonstrate how to operate these entities
Note that when we're editing relationships, we should keep both side in sync. E.g. When we add a new relationship between one book and one person, we should add reference in both side. Or, the graph will be in an inconsistent state and Neo4J does not know how to deal with it.